video_analyzer/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-filter==23.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
ffmpeg-python==0.2.0
openai-whisper==20231117
python-magic==0.4.27
celery==5.3.4
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0 